finish migration of methods to mystdocutils and mystsphinx
